refactor: split context/assembler.rs into focused modules
Split the 1649-line src/context/assembler.rs into five focused modules,
all under the 800-line threshold from AGENTS.md:

- assembler.rs (703 lines, down from 1649): ContextAssembler public API,
  budget helpers, and tests
- memory_selection.rs (595 lines): memory_selection() pipeline,
  retrieval candidates, JSON parsing helpers
- assembly_view.rs (203 lines): assemble_context_view()
- compaction_view.rs (113 lines): compaction_source_entries()
- retrieval_view.rs (74 lines): retrieval_source_entries()

No behavior changes. All public re-exports preserved through mod.rs.
